requirements.txt
cbf
aiohttp
pillow
